+++ plugins/postcss-custom-properties/src/is-var-function.ts
@@ -8,4 +8,4 @@ export function isVarFunction(node: Node): node is FunctionNode {
 	return node.type === 'function' && IS_VAR_FUNCTION_REGEX.test(node.value) && Object(node.nodes).length > 0;
 }
 
-export const HAS_VAR_FUNCTION_REGEX = /var\(/i;
+export const HAS_VAR_FUNCTION_REGEX = /\bvar\(/i;