+++ plugins/postcss-gradients-interpolation-method/src/index.ts
@@ -1,8 +1,7 @@
 import postcssProgressiveCustomProperties from '@csstools/postcss-progressive-custom-properties';
 import type { PluginCreator } from 'postcss';
 import { GRADIENT_FUNCTION_REGEX } from './is-gradient';
-import { hasFallback } from './has-fallback-decl';
-import { hasSupportsAtRuleAncestor } from './has-supports-at-rule-ancestor';
+import { hasFallback, hasSupportsAtRuleAncestor } from '@csstools/utilities';
 import { isFunctionNode, parseCommaSeparatedListOfComponentValues, replaceComponentValues, stringify } from '@csstools/css-parser-algorithms';
 import { modifyGradientFunctionComponentValues } from './modify-gradient-component-values';
 import { tokenize } from '@csstools/css-tokenizer';
@@ -24,7 +23,7 @@ const basePlugin: PluginCreator<basePluginOptions> = (opts?: basePluginOptions) 
 				return;
 			}
 
-			if (hasSupportsAtRuleAncestor(decl)) {
+			if (hasSupportsAtRuleAncestor(decl, GRADIENT_FUNCTION_REGEX)) {
 				return;
 			}